src/DictAugmentTransformers/requirements.txt
pandas
numpy
jsonlib2
python-dotenv
requests
regex
uuid
